+++ src/test/suite/kafka-file/languageservice/completionProperties.test.ts
@@ -569,6 +569,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 0))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
+                        range: range(position(1, 0), position(1, 0))
                     }
                 ]
             });
@@ -605,6 +610,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 1))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
+                        range: range(position(1, 0), position(1, 1))
                     }
                 ]
             });
@@ -639,6 +649,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        range: range(position(1, 0), position(1, 5))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
                         range: range(position(1, 0), position(1, 5))
                     }
                 ]
@@ -675,6 +690,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
+                        range: range(position(1, 0), position(1, 6))
                     }
                 ]
             });
@@ -710,6 +730,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
+                        range: range(position(1, 0), position(1, 6))
                     }
                 ]
             });
@@ -744,6 +769,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        range: range(position(1, 0), position(1, 11))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
                         range: range(position(1, 0), position(1, 11))
                     }
                 ]
@@ -781,6 +811,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
                         insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        range: range(position(2, 0), position(2, 5))
+                    },
+                    {
+                        label: 'every', kind: CompletionItemKind.Property,
+                        insertText: 'every: ${1|3s,5m,1h|}',
                         range: range(position(2, 0), position(2, 5))
                     }
                 ]